Conflicts:
	KiloLiters/.gradle/1.10/taskArtifacts/cache.properties.lock
	KiloLiters/app/src/main/res/layout/view_inserimento_dati.xml

+++ KiloLiters/app/src/main/java/it/beer2beer/kiloliters/DatabaseAdapter.java
@@ -77,7 +77,7 @@ public class DatabaseAdapter {
         DBHelper.close();
     }
 
-    public long insertRefuel(String timestamp, double chilometri,
+    public long insertRefuel(String timestamp, int chilometri,
                              double prezzo, double litri, double importo,
                              String distributore, String citta, String descrizione){
